Refactor zone transactions to always use versioned CoW code.
